requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.4
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.4.7
daphne==3.0.2
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-cors-headers==3.6.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.0.1
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
ipaddr==2.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
msgpack==1.0.2
packaging==20.3
pep517==0.8.2
progress==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.4.0